requirements.txt
asgiref==3.3.1
autopep8==1.5.5
Django==3.1.7
flake8==3.8.4
gunicorn==20.0.4
mccabe==0.6.1
Pillow==8.1.0
psycopg2-binary==2.8.6
pycodestyle==2.6.0
pyflakes==2.2.0
python-dotenv==0.15.0
pytz==2021.1
sorl-thumbnail==12.7.0
sqlparse==0.4.1
toml==0.10.2

numpy==1.20.1
names==0.3.0
djangorestframework==3.12.2